refactor : package quiz infra classes QUZ-47
