--- search-service/search-infra/src/main/kotlin/com/grepp/quizy/search/infra/quiz/UserAnswerLogDocument.kt
@@ -1,5 +1,0 @@
-package com.grepp.quizy.search.infra.quiz
-
-
-class UserAnswerLogDocument {
-}
