+++ search-service/search-infra/src/main/kotlin/com/grepp/quizy/search/infra/quiz/document/QuizDocumentVO.kt
@@ -1,4 +1,4 @@
-package com.grepp.quizy.search.infra.quiz
+package com.grepp.quizy.search.infra.quiz.document
 
 sealed interface QuizDocumentVo
