+++ search-service/search-infra/src/main/kotlin/com/grepp/quizy/search/infra/quiz/repository/QuizElasticRepository.kt
@@ -1,5 +1,6 @@
-package com.grepp.quizy.search.infra.quiz
+package com.grepp.quizy.search.infra.quiz.repository
 
+import com.grepp.quizy.search.infra.quiz.document.QuizDocument
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 
 interface QuizElasticRepository : ElasticsearchRepository<QuizDocument, Long>, CustomQuizSearchRepository
